Adding the eventListener back
Ok, I have a confession. I deleted the eventListener because my linter
was throwing a weird error saying a comma was needed where it surely
wasn't. So I deleted it. But for whatever reason now this isn't an issue. So what was the
issue? A misplaced closing parenthesis or bracket? Who knows!