ERD support & graphs, migrations
In an effort to populate the Landlord model with accurate data, I've
purged the landlords table of its empty relations. I've also purged the
properties table of the index from the landlords table; the index was
ambiguous and didn't accurately reflect unique landlords (the number of
landlords will be, as a rule, fewer than the total records in the properties table.
It's safe to assume there will be landlords who own more than one
property. We've already found this to be the case within the data).

ERD gem added; this is useful for seeing relationships between Models.