Package updates, seeds, validations
Instead of using a Rakefile task to seed the data, I've begun migrating
to seeds.rb in the db folder.

I added a few queries for ActiveRecord to the README.

The Properties list in properties/index.html.erb is now a table.

Properties model updated with a few validations.

Geocoder gem added.

A migration that added a foreign key to Landlord was deleted.