No infinite scrolling yet…but…
the foundation is all here.

`hooks` folder added. I had some trouble specifying the types for a hook that
tracks where the cursor is on the page. My choice to use
`React.RefObject<Element>` was inspired by [this
useIntersectionObserver](https://usehooks-ts.com/react-hook/use-intersection-observer)
component.

useInfiniteQuery added.

I edited the gitigonre to make an exception for react-table types.
These are not the official types. However, react-table does currently specify the types of its components, their prop and return
types, are clear and distinct in the API documentation.