useInfiniteQuery + eventListeners
useEventListener hook added…since essentially the infinite scroll
functionality is listening for when the user reaches the bottom. There
are a few ways to do this. I still haven't figured out the corresponding
changes in the properties_controller.

Cleaned up the `Properties` component; removed unused imports. Deleted
logic around react-table, since react-table appears to be a bit of a
mess at the moment. Maybe when they transition to hooks and TypeScript
fully it will be a better library.