SearchBar component was preventing the Properties
…component from updating; that has been changed. Event types specificed
inside the bootstrap component.

config/application.rb – explicitly telling Rails to function as an API,
since this is what I'm doing already.

Tried getting tests to work…toward that end I generated a few Address
fixtures. I am not sure if the tests are failing because I started this
app as a whole Rails app but then transitioned to only using it as an
API. A few systems tests were generated as scaffold…