Updates to controllers and models
I intend to use the Address model as an abstract association between
Landlord and Property.So Address is a middle term, but I still haven't
worked out the associations and their directions. Originally I geocoded
the concencatenated string, without accounting for municipality. I
manually updated the records. What I should have done was captured the
municapiality code from the csv filename.
 The municipality can be captured from the filename of the csv.
There should be a method or a dictionary for capturing and returning the translated value of the filename's municipality code.