Jbuilder, tests, and controller updates
JBuilder added to Properties and Landlords.

Address moved into `api/v1` folder where the other models are.

A join-table created after I noticed the backtrace of `rails erd` was showing an error
because it was missing.

Fixtures added for tests.

`Params` was incorrectly spelled in the `Properties` controller; not
anymore!